requirements.txt
opencv-python
Pillow
deepface
tf-keras
mediapipe
